refactor: :recycle: removed AddressDerivation dummy implementation and replaced it with mock
