+++ state-chain/pallets/cf-ingress/src/lib.rs
@@ -6,8 +6,6 @@
 // This way intents and intent ids align per chain, which makes sense given they act as an index to
 // the respective address generation function.
 
-use sp_std::str::FromStr;
-
 use cf_primitives::{ForeignChainAddress, ForeignChainAsset, IntentId};
 use cf_traits::{liquidity::LpProvisioningApi, AddressDerivationApi, FlipBalance, IngressApi};
 
@@ -17,7 +15,6 @@ mod benchmarking;
 pub mod weights;
 pub use weights::WeightInfo;
 
-use frame_support::sp_runtime::app_crypto::sp_core::H160;
 pub use pallet::*;
 
 #[frame_support::pallet]
@@ -183,20 +180,5 @@ impl<T: Config> IngressApi for Pallet<T> {
 		Self::deposit_event(Event::StartWitnessing { ingress_address, ingress_asset });
 
 		(intent_id, ingress_address)
-	}
-}
-
-pub struct KylesTestnetAddress;
-
-impl AddressDerivationApi for KylesTestnetAddress {
-	fn generate_address(
-		_ingress_asset: ForeignChainAsset,
-		_intent_id: IntentId,
-	) -> ForeignChainAddress {
-		ForeignChainAddress::Eth(
-			H160::from_str("F29aB9EbDb481BE48b80699758e6e9a3DBD609C6")
-				.unwrap()
-				.to_fixed_bytes(),
-		)
 	}
 }
